+++ userspace/falco/grpc_server.h
@@ -25,6 +25,7 @@ namespace falco
 {
 namespace grpc
 {
+
 class server : public server_impl
 {
 public:
@@ -50,6 +51,9 @@ public:
 	output::service::AsyncService m_svc;
 	std::unique_ptr<::grpc::ServerCompletionQueue> m_completion_queue;
 
+	// version::service::AsyncService m_version_svc;
+	// std::unique_ptr<::grpc::ServerCompletionQueue> m_version_completion_queue;
+
 private:
 	std::string m_server_addr;
 	int m_threadiness;
@@ -61,79 +65,5 @@ private:
 	std::vector<std::thread> m_threads;
 };
 
-class request_context_base
-{
-public:
-	request_context_base() = default;
-	~request_context_base() = default;
-
-	std::unique_ptr<::grpc::ServerContext> m_srv_ctx;
-	enum : char
-	{
-		UNKNOWN = 0,
-		REQUEST,
-		WRITE,
-		FINISH
-	} m_state = UNKNOWN;
-	virtual void start(server* srv) = 0;
-	virtual void process(server* srv) = 0;
-	virtual void end(server* srv, bool isError) = 0;
-};
-
-// The responsibility of `request_stream_context` template class
-// is to handle streaming responses.
-template<class Request, class Response>
-class request_stream_context : public request_context_base
-{
-public:
-	request_stream_context():
-		m_process_func(nullptr),
-		m_request_func(nullptr){};
-	~request_stream_context() = default;
-
-	// Pointer to function that does actual processing
-	void (server::*m_process_func)(const stream_context&, const Request&, Response&);
-
-	// fixme(leodido) > why output::service:: ... ?
-	// Pointer to function that requests the system to start processing given requests
-	void (output::service::AsyncService::*m_request_func)(::grpc::ServerContext*, Request*, ::grpc::ServerAsyncWriter<Response>*, ::grpc::CompletionQueue*, ::grpc::ServerCompletionQueue*, void*);
-
-	void start(server* srv);
-	void process(server* srv);
-	void end(server* srv, bool isError);
-
-private:
-	std::unique_ptr<::grpc::ServerAsyncWriter<Response>> m_res_writer;
-	std::unique_ptr<stream_context> m_stream_ctx;
-	Request m_req;
-};
-
-// The responsibility of `request_context` template class
-// is to handle unary responses.
-template<class Service, class Request, class Response>
-class request_context : public request_context_base
-{
-public:
-	request_context():
-		m_process_func(nullptr),
-		m_request_func(nullptr){};
-	~request_context() = default;
-
-	// Pointer to function that does actual processing
-	void (server::*m_process_func)(const context&, const Request&, Response&);
-
-	// fixme(leodido) > why output::service:: ... ?
-	// Pointer to function that requests the system to start processing given requests
-	void (Service::AsyncService::*m_request_func)(::grpc::ServerContext*, Request*, ::grpc::ServerAsyncWriter<Response>*, ::grpc::CompletionQueue*, ::grpc::ServerCompletionQueue*, void*);
-
-	void start(server* srv);
-	void process(server* srv);
-	void end(server* srv, bool isError);
-
-private:
-	// todo(leodido) > factorize these two into tbe base class?
-	std::unique_ptr<::grpc::ServerAsyncWriter<Response>> m_res_writer;
-	Request m_req;
-};
 } // namespace grpc
 } // namespace falco
